+++ Ryujinx.HLE/HOS/Applets/SoftwareKeyboard/SoftwareKeyboardState.cs
@@ -1,5 +1,8 @@
 ﻿namespace Ryujinx.HLE.HOS.Applets.SoftwareKeyboard
 {
+    /// <summary>
+    /// Identifies the software keyboard state.
+    /// </summary>
     internal enum SoftwareKeyboardState
     {
         /// <summary>
